Improve fetching blog function
