+++ apps/engineeringblog/components/Article/index.ts
@@ -1,5 +1,5 @@
-import ArtilceList from "./ArtilceList";
+import ArticleList from "./ArticleList";
 import ArticleCard from "./ArticleCard";
 import ArticleContent from "./ArticleContent";
 
-export { ArtilceList, ArticleCard, ArticleContent };
+export { ArticleList, ArticleCard, ArticleContent };
